Update: ItemInfo CRUD API 일괄 수정
- 기본 요청 경로 : /api/v1/info/items
- 기존에 사용하던 시간에 등장하는 모든 아이템들을 한번에 저장하는 API 더이상 사용하지 않을꺼 같아서 Summary에 Deprecated라고 적어둠
- Read 요청 이외에는 전부 유튜버 권한이 있어야 요청 가능하도록 함.
- 수정, 삭제는 작성자만 가능

+++ src/main/java/com/clovi/service/item/ItemInfoService.java
@@ -5,11 +5,14 @@ import com.clovi.domain.item.Item;
 import com.clovi.domain.item.ItemInfo;
 import com.clovi.domain.item.ItemColor;
 import com.clovi.domain.item.ItemSize;
+import com.clovi.domain.user.Member;
 import com.clovi.dto.requests.item.ItemInfoCreateRequest;
 import com.clovi.dto.requests.item.ItemInfoDeleteRequest;
 import com.clovi.dto.requests.item.ItemInfoUpdateRequest;
 import com.clovi.dto.response.ItemResponseDto;
 import com.clovi.exception.ResourceNotFoundException;
+import com.clovi.exception.auth.NoPermissionDeleteException;
+import com.clovi.exception.auth.NoPermissionUpdateException;
 import com.clovi.repository.Category.CategoryRepository;
 import com.clovi.repository.Item.ItemRepository;
 import com.clovi.repository.Item.ItemInfoRepository;
@@ -44,17 +47,27 @@ public class ItemInfoService {
     return itemInfo;
   }
 
-  // Used by ItemController
-  public ItemResponseDto getItemById(Long itemId) {
-    ItemInfo itemInfo = itemInfoRepository.findById(itemId).orElseThrow(() -> new ResourceNotFoundException("아이템",itemId));
+  // Used by ItemInfoController
+  public ItemResponseDto getItemByIdV1(Long itemInfoId) {
+    ItemInfo itemInfo = itemInfoRepository.findByIdAndDeletedIsFalse(itemInfoId).orElseThrow(() -> new ResourceNotFoundException("아이템",itemInfoId));
     List<Item> item = itemRepository.findAllByItemInfoIdAndDeletedIsFalse(itemInfo.getId());
+    if(item.size() == 0 ){
+      return new ItemResponseDto(itemInfo);
+    }
     return new ItemResponseDto(itemInfo,item.get(0));
   }
-
+  //Item 페이지에서 color 랑 size 다른것들 보여줄때 사용할 수 도 있을 거 같음 .
+  public ItemResponseDto getItemByIdV2(Long itemInfoId) {
+    ItemInfo itemInfo = itemInfoRepository.findById(itemInfoId).orElseThrow(() -> new ResourceNotFoundException("아이템",itemInfoId));
+    List<Item> item = itemRepository.findAllByItemInfoIdAndDeletedIsFalse(itemInfo.getId());
+    return new ItemResponseDto(itemInfo,item);
+  }
   // 아래의 세 메소드 모두 유저 찾아서 권한 확인하는 부분 추가해야함.
   @Transactional
-  public Long create(ItemInfoCreateRequest itemInfoCreateRequest, Long userId){
+  public Long create(ItemInfoCreateRequest itemInfoCreateRequest, Member member){
     Long categoryId = itemInfoCreateRequest.getCategoryId();
+
+    Long userId = member.getId();
 
     Category category = categoryRepository.findByIdAndDeletedIsFalse(categoryId).orElseThrow(() -> new ResourceNotFoundException("category",categoryId));
 
@@ -65,11 +78,16 @@ public class ItemInfoService {
     return saved.getId();
   }
   @Transactional
-  public Long update(ItemInfoUpdateRequest itemInfoUpdateRequest, Long userId){
-    Long itemId = itemInfoUpdateRequest.getItemId();
+  public Long update(ItemInfoUpdateRequest itemInfoUpdateRequest, Long itemInfoId,  Member member){
     Long categoryId = itemInfoUpdateRequest.getCategoryId();
+    Long userId = member.getId();
 
-    ItemInfo findItemInfo = itemInfoRepository.findByIdAndDeletedIsFalse(itemId).orElseThrow(() -> new ResourceNotFoundException("Item", itemId));
+
+    ItemInfo findItemInfo = itemInfoRepository.findByIdAndDeletedIsFalse(itemInfoId).orElseThrow(() -> new ResourceNotFoundException("Item", itemInfoId));
+
+    if(findItemInfo.isNotCreatedBy(userId)){
+      throw new NoPermissionUpdateException();
+    }
 
     Category category = categoryRepository.findByIdAndDeletedIsFalse(categoryId).orElseThrow(() -> new ResourceNotFoundException("category",categoryId));
 
@@ -81,10 +99,15 @@ public class ItemInfoService {
   }
 
   @Transactional
-  public void delete(ItemInfoDeleteRequest itemInfoDeleteRequest, Long userId){
-    Long itemId = itemInfoDeleteRequest.getItemId();
+  public void delete(Long itemInfoId,  Member member){
+    Long userId = member.getId();
 
-    ItemInfo findItemInfo = itemInfoRepository.findByIdAndDeletedIsFalse(itemId).orElseThrow(() -> new ResourceNotFoundException("Item", itemId));
+
+    ItemInfo findItemInfo = itemInfoRepository.findByIdAndDeletedIsFalse(itemInfoId).orElseThrow(() -> new ResourceNotFoundException("Item", itemInfoId));
+
+    if(findItemInfo.isNotCreatedBy(userId)){
+      throw new NoPermissionDeleteException();
+    }
 
     findItemInfo.delete();
